refactor(api_fastapi): extract asset_event datamodels from asset
